Исправление цикла for на for each, исправление byte на duoble, исправление расчётов

+++ src/main/java/model/Discountable.java
@@ -2,5 +2,5 @@ package model;
 
 public interface Discountable {
 
-    byte getDiscount();
+    double getDiscount();
 }
